+++ linepy/__init__.py
@@ -1,14 +1,13 @@
 from .client import LineClient
 from .channel import LineChannel
-from .call import LineCall
 from .poll import LinePoll
 from akad.ttypes import OpType
 
 __copyright__       = 'Copyright 2017 by Fadhiil Rachman'
-__version__         = '1.9.5'
+__version__         = '2.0.0'
 __license__         = 'BSD-3-Clause'
 __author__          = 'Fadhiil Rachman'
 __author_email__    = 'fadhiilrachman@gmail.com'
 __url__             = 'http://github.com/fadhiilrachman/line-py'
 
-__all__ = ['LineClient', 'LineChannel', 'LineCall', 'LinePoll', 'OpType']
+__all__ = ['LineClient', 'LineChannel', 'LinePoll', 'OpType']