+++ linepy/callback.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-class LineCallback(object):
+class Callback(object):
 
     def __init__(self, callback):
         self.callback = callback
@@ -9,9 +9,9 @@ class LineCallback(object):
 
     def QrUrl(self, url, showQr=True):
         if showQr:
-            notice='or scan this QR '
+            notice="or scan this QR "
         else:
-            notice=''
+            notice=""
         self.callback("Open this link " + notice + "on your LINE for smartphone in 2 minutes\n" + url)
         if showQr:
             try: