+++ linepy/session.py
@@ -1,13 +1,25 @@
 # -*- coding: utf-8 -*-
 from thrift.transport import THttpClient
 from thrift.protocol import TCompactProtocol
-from akad import TalkService, ChannelService, CallService, SquareService
+from akad import AuthService, TalkService, ChannelService, CallService, SquareService
 
 class LineSession:
 
     def __init__(self, url, headers, path=''):
         self.host = url + path
         self.headers = headers
+
+    def Auth(self, isopen=True):
+        self.transport = THttpClient.THttpClient(self.host)
+        self.transport.setCustomHeaders(self.headers)
+        
+        self.protocol = TCompactProtocol.TCompactProtocol(self.transport)
+        self._auth  = AuthService.Client(self.protocol)
+        
+        if isopen:
+            self.transport.open()
+
+        return self._auth
 
     def Talk(self, isopen=True):
         self.transport = THttpClient.THttpClient(self.host)