+++ linepy/talk.py
@@ -1,0 +1,354 @@
+# -*- coding: utf-8 -*-
+from akad.ttypes import Message
+from random import randint
+
+import json, ntpath
+
+def loggedIn(func):
+    def checkLogin(*args, **kwargs):
+        if args[0].isLogin:
+            return func(*args, **kwargs)
+        else:
+            args[0].callback.other("You want to call the function, you must login to LINE")
+    return checkLogin
+
+class LineTalk(object):
+    isLogin = False
+
+    def __init__(self):
+        self.isLogin = True
+
+    """User"""
+
+    @loggedIn
+    def getProfile(self):
+        return self.talk.getProfile()
+
+    @loggedIn
+    def getSettings(self):
+        return self.talk.getSettings()
+
+    @loggedIn
+    def getUserTicket(self):
+        return self.talk.getUserTicket()
+
+    @loggedIn
+    def updateProfile(self, profileObject):
+        return self.talk.updateProfile(0, profileObject)
+
+    @loggedIn
+    def updateSettings(self, settingObject):
+        return self.talk.updateSettings(0, settingObject)
+
+    @loggedIn
+    def updateProfileAttribute(self, attrId, value):
+        return self.talk.updateProfileAttribute(0, attrId, value)
+
+    """Operation"""
+
+    @loggedIn
+    def fetchOperation(self, revision, count):
+        return self.talk.fetchOperations(revision, count)
+
+    @loggedIn
+    def getLastOpRevision(self):
+        return self.talk.getLastOpRevision()
+
+    """Message"""
+
+    @loggedIn
+    def sendMessage(self, to, text, contentMetadata={}, contentType=0):
+        msg = Message()
+        msg.to, msg._from = to, self.profile.mid
+        msg.text = text
+        msg.contentType, msg.contentMetadata = contentType, contentMetadata
+        if to not in self._messageReq:
+            self._messageReq[to] = -1
+        self._messageReq[to] += 1
+        return self.talk.sendMessage(self._messageReq[to], msg)
+    
+    """ Usage:
+        @to Integer
+        @text String
+        @dataMid List of user Mid
+    """
+    @loggedIn
+    def sendMessageWithMention(self, to, text='', dataMid=[]):
+        arr = []
+        list_text=''
+        if '[list]' in text.lower():
+            i=0
+            for l in dataMid:
+                list_text+='\n@[list-'+str(i)+']'
+                i=i+1
+            text=text.replace('[list]', list_text)
+        elif '[list-' in text.lower():
+            text=text
+        else:
+            i=0
+            for l in dataMid:
+                list_text+=' @[list-'+str(i)+']'
+                i=i+1
+            text=text+list_text
+        i=0
+        for l in dataMid:
+            mid=l
+            name='@[list-'+str(i)+']'
+            ln_text=text.replace('\n',' ')
+            if ln_text.find(name):
+                line_s=int(ln_text.index(name))
+                line_e=(int(line_s)+int(len(name)))
+            arrData={'S': str(line_s), 'E': str(line_e), 'M': mid}
+            arr.append(arrData)
+            i=i+1
+        contentMetadata={'MENTION':str('{"MENTIONEES":' + json.dumps(arr).replace(' ','') + '}')}
+        return self.sendMessage(to, text, contentMetadata)
+
+    @loggedIn
+    def sendSticker(self, to, packageId, stickerId):
+        contentMetadata = {
+            'STKVER': '100',
+            'STKPKGID': packageId,
+            'STKID': stickerId
+        }
+        return self.sendMessage(to, '', contentMetadata, 7)
+        
+    @loggedIn
+    def sendContact(self, to, mid):
+        contentMetadata = {'mid': mid}
+        return self.sendMessage(to, '', contentMetadata, 13)
+
+    @loggedIn
+    def sendGift(self, productId, productType):
+        if productType not in ['theme','sticker']:
+            raise Exception('Invalid productType value')
+        contentMetadata = {
+            'MSGTPL': str(randint(0, 12)),
+            'PRDTYPE': productType.upper(),
+            'STKPKGID' if productType == 'sticker' else 'PRDID': productId
+        }
+        return self.sendMessage(to, '', contentMetadata, 9)
+
+    @loggedIn
+    def removeMessage(self, messageId):
+        return self.talk.removeMessage(messageId)
+        
+    @loggedIn
+    def removeAllMessages(self, lastMessageId):
+        return self.talk.removeAllMessages(0, lastMessageId)
+        
+    @loggedIn
+    def sendChatChecked(self, consumer, messageId):
+        return self.talk.sendChatChecked(0, consumer, messageId)
+
+    @loggedIn
+    def sendEvent(self, messageObject):
+        return self.talk.sendEvent(0, messageObject)
+
+    @loggedIn
+    def getLastReadMessageIds(self, chatId):
+        return self.talk.getLastReadMessageIds(0,chatId)
+
+    """Object"""
+
+    @loggedIn
+    def sendImage(self, to, path):
+        objectId = self.sendMessage(to=to, text=None, contentType = 1).id
+        return self.uploadObjTalk(path=path, type='image', returnAs='bool', objId=objectId)
+
+    @loggedIn
+    def sendImageWithURL(self, to, url):
+        path = self.downloadFileURL(url, 'path')
+        return self.sendImage(to, path)
+
+    @loggedIn
+    def sendGIF(self, to, path):
+        return self.uploadObjTalk(path=path, type='gif', returnAs='bool', to=to)
+
+    @loggedIn
+    def sendGIFWithURL(self, to, url):
+        path = self.downloadFileURL(url, 'path')
+        return self.sendGIF(to, path)
+
+    @loggedIn
+    def sendVideo(self, to, path):
+        objectId = self.sendMessage(to=to, text=None, contentMetadata={'VIDLEN': '60000','DURATION': '60000'}, contentType = 2).id
+        return self.uploadObjTalk(path=path, type='video', returnAs='bool', objId=objectId)
+
+    @loggedIn
+    def sendVideoWithURL(self, to, url):
+        path = self.downloadFileURL(url, 'path')
+        return self.sendVideo(to, path)
+
+    @loggedIn
+    def sendAudio(self, to, path):
+        objectId = self.sendMessage(to=to, text=None, contentType = 3).id
+        return self.uploadObjTalk(path=path, type='audio', returnAs='bool', objId=objectId)
+
+    @loggedIn
+    def sendAudioWithURL(self, to, url):
+        path = self.downloadFileURL(url, 'path')
+        return self.sendAudio(to, path)
+
+    @loggedIn
+    def sendFile(self, to, path, file_name=''):
+        if file_name == '':
+            file_name = ntpath.basename(path)
+        file_size = len(open(path, 'rb').read())
+        objectId = self.sendMessage(to=to, text=None, contentMetadata={'FILE_NAME': str(file_name),'FILE_SIZE': str(file_size)}, contentType = 14).id
+        return self.uploadObjTalk(path=path, type='file', returnAs='bool', objId=objectId)
+
+    @loggedIn
+    def sendFileWithURL(self, to, url, fileName=''):
+        path = self.downloadFileURL(url, 'path')
+        return self.sendFile(to, path, fileName)
+
+    """Contact"""
+        
+    @loggedIn
+    def blockContact(self, mid):
+        return self.talk.blockContact(0, mid)
+
+    @loggedIn
+    def unblockContact(self, mid):
+        return self.talk.unblockContact(0, mid)
+
+    @loggedIn
+    def findAndAddContactsByMid(self, mid):
+        return self.talk.findAndAddContactsByMid(0, mid)
+
+    @loggedIn
+    def findAndAddContactsByUserid(self, userid):
+        return self.talk.findAndAddContactsByUserid(0, userid)
+
+    @loggedIn
+    def findContactsByUserid(self, userid):
+        return self.talk.findContactByUserid(userid)
+
+    @loggedIn
+    def findContactByTicket(self, ticketId):
+        return self.talk.findContactByUserTicket(ticketId)
+
+    @loggedIn
+    def getAllContactIds(self):
+        return self.talk.getAllContactIds()
+
+    @loggedIn
+    def getBlockedContactIds(self):
+        return self.talk.getBlockedContactIds()
+
+    @loggedIn
+    def getContact(self, mid):
+        return self.talk.getContact(mid)
+
+    @loggedIn
+    def getContacts(self, midlist):
+        return self.talk.getContacts(midlist)
+
+    @loggedIn
+    def getFavoriteMids(self):
+        return self.talk.getFavoriteMids()
+
+    @loggedIn
+    def getHiddenContactMids(self):
+        return self.talk.getHiddenContactMids()
+
+    @loggedIn
+    def reissueUserTicket(self, expirationTime=100, maxUseCount=100):
+        return self.talk.reissueUserTicket(expirationTime, maxUseCount)
+    
+    @loggedIn
+    def cloneContactProfile(self, mid):
+        contact = self.getContact(mid)
+        profile = self.profile
+        profile.displayName = contact.displayName
+        profile.statusMessage = contact.statusMessage
+        profile.pictureStatus = contact.pictureStatus
+        self.updateProfileAttribute(8, profile.pictureStatus)
+        return self.updateProfile(profile)
+
+    """Group"""
+    
+    @loggedIn
+    def findGroupByTicket(self, ticketId):
+        return self.talk.findGroupByTicket(ticketId)
+
+    @loggedIn
+    def acceptGroupInvitation(self, groupId):
+        return self.talk.acceptGroupInvitation(0, groupId)
+
+    @loggedIn
+    def acceptGroupInvitationByTicket(self, groupId, ticketId):
+        return self.talk.acceptGroupInvitationByTicket(0, groupId, ticketId)
+
+    @loggedIn
+    def cancelGroupInvitation(self, groupId, contactIds):
+        return self.talk.cancelGroupInvitation(0, groupId, contactIds)
+
+    @loggedIn
+    def createGroup(self, name, midlist):
+        return self.talk.createGroup(0, name, midlist)
+
+    @loggedIn
+    def getGroup(self, groupId):
+        return self.talk.getGroup(groupId)
+
+    @loggedIn
+    def getGroups(self, groupIds):
+        return self.talk.getGroups(groupIds)
+
+    @loggedIn
+    def getGroupIdsInvited(self):
+        return self.talk.getGroupIdsInvited()
+
+    @loggedIn
+    def getGroupIdsJoined(self):
+        return self.talk.getGroupIdsJoined()
+
+    @loggedIn
+    def inviteIntoGroup(self, groupId, midlist):
+        return self.talk.inviteIntoGroup(0, groupId, midlist)
+
+    @loggedIn
+    def kickoutFromGroup(self, groupId, midlist):
+        return self.talk.kickoutFromGroup(0, groupId, midlist)
+
+    @loggedIn
+    def leaveGroup(self, groupId):
+        return self.talk.leaveGroup(0, groupId)
+
+    @loggedIn
+    def rejectGroupInvitation(self, groupId):
+        return self.talk.rejectGroupInvitation(0, groupId)
+
+    @loggedIn
+    def reissueGroupTicket(self, groupId):
+        return self.talk.reissueGroupTicket(groupId)
+
+    @loggedIn
+    def updateGroup(self, groupObject):
+        return self.talk.updateGroup(0, groupObject)
+
+    """Room"""
+
+    @loggedIn
+    def createRoom(self, midlist):
+        return self.talk.createRoom(0, midlist)
+
+    @loggedIn
+    def getRoom(self, roomId):
+        return self.talk.getRoom(roomId)
+
+    @loggedIn
+    def inviteIntoRoom(self, roomId, midlist):
+        return self.talk.inviteIntoRoom(0, roomId, midlist)
+
+    @loggedIn
+    def leaveRoom(self, roomId):
+        return self.talk.leaveRoom(0, roomId)
+
+    """Call"""
+        
+    @loggedIn
+    def acquireCallTalkRoute(self, to):
+        return self.talk.acquireCallRoute(to)